fix: code splitting (ShopOwnerDashboard)
