+++ pages/dashboard.tsx
@@ -1,11 +1,11 @@
+import ShopOwnerDashboard from '@modules/dashboard/component/dashboard';
 import React from 'react';
 // import withAuth from '../helpers/withAuth';
 import NavDashBoard from '../modules/dashboard/component/Navbar';
-import ShopOwnerDashboard from '../modules/dashboard/component/ShopOwnerDashboard';
 
 function dashboard() {
   return (
-    <div className="px-[50.5px] md:px-[101px] min-h-[100svh]">
+    <div className="px-[25px] md:px-[50px] lg:px-[75px] xl:px-[101px] min-h-[100svh]">
       <NavDashBoard active="dashboard" />
       <ShopOwnerDashboard />
     </div>
